+++ ZenUnitUtilsAndAssertionTests/Utils/Time/Mock/StopwatchMock.h
@@ -4,5 +4,5 @@ class StopwatchMock : public Zen::Mock<ZenUnit::Stopwatch>
 {
 public:
    ZENMOCK_VOID0(Start)
-   ZENMOCK_NONVOID0(unsigned, Stop)
+   ZENMOCK_NONVOID0(long long, Stop)
 };
